fix: zustand로 로그인 상태관리
